add a '^' to deps in package.json
this will make it so if the thing consuming instructure-icons
has other things that require these same dependencies
but, for example, require a newer version than what we have here
they and this will use the same version of that dependency.
It will make it so the webpack bundle sizes in quizzes
and canvas-lms are smaller.

Change-Id: Iaeef168fb0e6740ca57a81bfcb3f619c50dbcd87
Reviewed-on: https://gerrit.instructure.com/117233
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jennifer Stern <jstern@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>